+++ mtg_proxy_printer/decklist_parser/re_parsers.py
@@ -22,6 +22,9 @@ from PyQt5.QtCore import QObject
 from mtg_proxy_printer.decklist_parser.common import ParsedDeck, ParserBase
 from mtg_proxy_printer.model.carddb import Card, CardDatabase, CardIdentificationData
 from mtg_proxy_printer.model.imagedb import ImageDatabase
+from mtg_proxy_printer.logger import get_logger
+logger = get_logger(__name__)
+del get_logger
 
 MatchType = typing.Dict[str, str]
 
@@ -50,6 +53,7 @@ class GenericRegularExpressionDeckParser(ParserBase):
         self.parser = regular_expression \
             if isinstance(regular_expression, re.Pattern) \
             else re.compile(regular_expression)
+        logger.info(f"Created {self.__class__.__name__} instance using RE '{regular_expression}'")
 
     def parse_deck_without_translation(self, deck_list: str,
                                        print_guessing: bool) -> ParsedDeck: